refactor: 예외 처리 기능 책임 분리
