+++ src/main/java/cart/controller/CartControllerAdvice.java
@@ -1,0 +1,23 @@
+package cart.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class CartControllerAdvice {
+    public static final String SERVER_ERROR = "server error";
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<String> handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
+        String errorMessage = e.getBindingResult().getAllErrors().get(0).getDefaultMessage();
+        return new ResponseEntity<>(errorMessage, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<String> handleServerException(Exception e) {
+        return new ResponseEntity<>(SERVER_ERROR, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+}
